+++ libcaf_core/caf/all.hpp
@@ -6,7 +6,6 @@
 
 #include "caf/abstract_actor.hpp"
 #include "caf/abstract_channel.hpp"
-#include "caf/abstract_group.hpp"
 #include "caf/actor.hpp"
 #include "caf/actor_addr.hpp"
 #include "caf/actor_clock.hpp"
@@ -45,7 +44,6 @@
 #include "caf/expected.hpp"
 #include "caf/extend.hpp"
 #include "caf/function_view.hpp"
-#include "caf/group.hpp"
 #include "caf/hash/fnv.hpp"
 #include "caf/init_global_meta_objects.hpp"
 #include "caf/local_actor.hpp"