+++ libcaf_core/caf/fwd.hpp
@@ -73,7 +73,6 @@ template <class...> class typed_response_promise;
 
 class [[nodiscard]] error;
 class abstract_actor;
-class abstract_group;
 class abstract_mailbox;
 class action;
 class actor;
@@ -104,8 +103,6 @@ class disposable;
 class event_based_actor;
 class execution_unit;
 class forwarding_actor_proxy;
-class group;
-class group_module;
 class hashed_node_id;
 class ipv4_address;
 class ipv4_endpoint;
@@ -153,7 +150,6 @@ class stateful_actor;
 
 struct down_msg;
 struct exit_msg;
-struct group_down_msg;
 struct illegal_message_element;
 struct invalid_actor_addr_t;
 struct invalid_actor_t;
@@ -360,7 +356,6 @@ class abstract_worker;
 class abstract_worker_hub;
 class disposer;
 class dynamic_message_data;
-class group_manager;
 class message_data;
 class private_thread;
 class stream_bridge;
@@ -384,7 +379,6 @@ using weak_actor_ptr = weak_intrusive_ptr<actor_control_block>;
 
 // -- intrusive pointer aliases ------------------------------------------------
 
-using group_module_ptr = intrusive_ptr<group_module>;
 using strong_actor_ptr = intrusive_ptr<actor_control_block>;
 
 // -- unique pointer aliases ---------------------------------------------------