+++ libcaf_core/caf/local_actor.hpp
@@ -5,7 +5,6 @@
 #pragma once
 
 #include "caf/abstract_actor.hpp"
-#include "caf/abstract_group.hpp"
 #include "caf/actor.hpp"
 #include "caf/actor_cast.hpp"
 #include "caf/actor_config.hpp"
@@ -132,55 +131,6 @@ public:
     detail::bool_token<spawnable> enabled;
     return eval_opts(Os, system().spawn_functor<unbound>(
                            enabled, cfg, fun, std::forward<Ts>(xs)...));
-  }
-
-  template <class T, spawn_options Os = no_spawn_options, class Groups,
-            class... Ts>
-  actor spawn_in_groups(const Groups& gs, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    return eval_opts(Os, system().spawn_class_in_groups<T, make_unbound(Os)>(
-                           cfg, gs.begin(), gs.end(), std::forward<Ts>(xs)...));
-  }
-
-  template <class T, spawn_options Os = no_spawn_options, class... Ts>
-  actor spawn_in_groups(std::initializer_list<group> gs, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    return eval_opts(Os, system().spawn_class_in_groups<T, make_unbound(Os)>(
-                           cfg, gs.begin(), gs.end(), std::forward<Ts>(xs)...));
-  }
-
-  template <class T, spawn_options Os = no_spawn_options, class... Ts>
-  actor spawn_in_group(const group& grp, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    auto first = &grp;
-    return eval_opts(Os, system().spawn_class_in_groups<T, make_unbound(Os)>(
-                           cfg, first, first + 1, std::forward<Ts>(xs)...));
-  }
-
-  template <spawn_options Os = no_spawn_options, class Groups, class F,
-            class... Ts>
-  actor spawn_in_groups(const Groups& gs, F fun, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    return eval_opts(Os, system().spawn_fun_in_groups<make_unbound(Os)>(
-                           cfg, gs.begin(), gs.end(), fun,
-                           std::forward<Ts>(xs)...));
-  }
-
-  template <spawn_options Os = no_spawn_options, class F, class... Ts>
-  actor spawn_in_groups(std::initializer_list<group> gs, F fun, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    return eval_opts(Os, system().spawn_fun_in_groups<make_unbound(Os)>(
-                           cfg, gs.begin(), gs.end(), fun,
-                           std::forward<Ts>(xs)...));
-  }
-
-  template <spawn_options Os = no_spawn_options, class F, class... Ts>
-  actor spawn_in_group(const group& grp, F fun, Ts&&... xs) {
-    actor_config cfg{context(), this};
-    auto first = &grp;
-    return eval_opts(Os,
-                     system().spawn_fun_in_groups<make_unbound(Os)>(
-                       cfg, first, first + 1, fun, std::forward<Ts>(xs)...));
   }
 
   // -- sending asynchronous messages ------------------------------------------