+++ libcaf_core/tests/legacy/mixin/sender.cpp
@@ -26,14 +26,12 @@ behavior testee_impl(event_based_actor* self) {
 }
 
 struct fixture : test_coordinator_fixture<> {
-  group grp;
   actor testee;
 
   std::string hello = "hello world";
 
   fixture() {
-    grp = sys.groups().anonymous();
-    testee = sys.spawn_in_group(grp, testee_impl);
+    testee = sys.spawn(testee_impl);
   }
 
   ~fixture() {
@@ -56,17 +54,6 @@ CAF_TEST(delayed actor messages receive responses) {
   expect((std::string), from(testee).to(self).with(hello));
 }
 
-CAF_TEST(delayed group message receive responses) {
-  self->delayed_send(grp, seconds(1), hello);
-  sched.trigger_timeout();
-  expect((std::string), from(self).to(testee).with(hello));
-  expect((std::string), from(testee).to(self).with(hello));
-  self->scheduled_send(grp, self->clock().now() + seconds(1), hello);
-  sched.trigger_timeout();
-  expect((std::string), from(self).to(testee).with(hello));
-  expect((std::string), from(testee).to(self).with(hello));
-}
-
 CAF_TEST(anonymous messages receive no response) {
   self->anon_send(testee, hello);
   expect((std::string), to(testee).with(hello));
@@ -76,14 +63,6 @@ CAF_TEST(anonymous messages receive no response) {
   expect((std::string), to(testee).with(hello));
   disallow((std::string), from(testee).to(self).with(hello));
   self->scheduled_anon_send(testee, self->clock().now() + seconds(1), hello);
-  sched.trigger_timeout();
-  expect((std::string), to(testee).with(hello));
-  disallow((std::string), from(testee).to(self).with(hello));
-  self->delayed_anon_send(grp, seconds(1), hello);
-  sched.trigger_timeout();
-  expect((std::string), to(testee).with(hello));
-  disallow((std::string), from(testee).to(self).with(hello));
-  self->scheduled_anon_send(grp, self->clock().now() + seconds(1), hello);
   sched.trigger_timeout();
   expect((std::string), to(testee).with(hello));
   disallow((std::string), from(testee).to(self).with(hello));